Plugin/requirements.txt
networkx
gurobipy
numpy
scipy
pyln-client>=0.12.1